Revert "Test : 테스트 코드 작성"
This reverts commit 00c7ec1d38835b25c5b33c3a26777c7e82d5ea2d.

+++ app/src/main/java/com/example/fitnutrijournal/data/model/Meal.kt
@@ -11,5 +11,5 @@ data class Meal(
     val date: String,
     val mealType: String, // "아침", "점심", "저녁", "간식"
     val dietFoodCode: String,
-    val quantity: Int // 섭취한 양(g)
+    val quantity: Float // 섭취한 양(g)
 )
